+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/analysis/AnalysisController.java
@@ -72,12 +72,16 @@ public class AnalysisController {
 	private static final Logger logger = LoggerFactory.getLogger(AnalysisController.class);
 	// PAGES
 	public static final Map<AnalysisType, String> PREVIEWS = ImmutableMap
-			.of(AnalysisType.PHYLOGENOMICS, "tree", AnalysisType.SISTR_TYPING, "sistr");
+			.of(AnalysisType.PHYLOGENOMICS, "tree", AnalysisType.SISTR_TYPING, "sistr",
+					AnalysisType.MLST_MENTALIST, "tree");
 	private static final String BASE = "analysis/";
 	public static final String PAGE_DETAILS_DIRECTORY = BASE + "details/";
 	public static final String PREVIEW_UNAVAILABLE = PAGE_DETAILS_DIRECTORY + "unavailable";
 	public static final String PAGE_ANALYSIS_LIST = "analyses/analyses";
 	public static final String PAGE_USER_ANALYSIS_OUPUTS = "analyses/user-analysis-outputs";
+
+	private static final String TREE_EXT = "newick";
+	private static final String EMPTY_TREE = "();";
 
 	/*
 	 * SERVICES
@@ -284,7 +288,7 @@ public class AnalysisController {
 		try {
 			if (submission.getAnalysisState()
 					.equals(AnalysisState.COMPLETED)) {
-				if (analysisType.equals(AnalysisType.PHYLOGENOMICS)) {
+				if (analysisType.equals(AnalysisType.PHYLOGENOMICS) || analysisType.equals(AnalysisType.MLST_MENTALIST)) {
 					tree(submission, model);
 				} else if (analysisType.equals(AnalysisType.SISTR_TYPING)) {
 					model.addAttribute("sistr", true);
@@ -344,16 +348,20 @@ public class AnalysisController {
 	 *
 	 * @param id {@link AnalysisSubmission} id
 	 * @return map of info about each {@link AnalysisOutputFile}
+	 * @throws IridaWorkflowNotFoundException if the specified workflow cannot be found.
 	 */
 	@RequestMapping(value = "/ajax/{id}/outputs", method = RequestMethod.GET)
 	@ResponseBody
-	public List<AnalysisOutputFileInfo> getOutputFilesInfo(@PathVariable Long id) {
+	public List<AnalysisOutputFileInfo> getOutputFilesInfo(@PathVariable Long id)
+			throws IridaWorkflowNotFoundException {
 		AnalysisSubmission submission = analysisSubmissionService.read(id);
 		Analysis analysis = submission.getAnalysis();
-		Set<String> outputNames = analysis.getAnalysisOutputFileNames();
+		final List<String> outputNames = workflowsService.getOutputNames(submission.getWorkflowId());
 		return outputNames.stream()
 				.map((outputName) -> getAnalysisOutputFileInfo(submission, analysis, outputName))
 				.filter(Objects::nonNull)
+				.filter(x -> x.getFileSizeBytes() > 0L)
+				.filter(x -> !(TREE_EXT.equals(x.getFileExt()) && EMPTY_TREE.equals(x.getFirstLine())))
 				.collect(Collectors.toList());
 	}
 
@@ -369,7 +377,7 @@ public class AnalysisController {
 			String outputName) {
 		final ImmutableSet<String> BLACKLIST_FILE_EXT = ImmutableSet.of("zip");
 		// set of file extensions for indicating whether the first line of the file should be read
-		final ImmutableSet<String> FILE_EXT_READ_FIRST_LINE = ImmutableSet.of("tsv", "txt", "tabular", "csv", "tab");
+		final ImmutableSet<String> FILE_EXT_READ_FIRST_LINE = ImmutableSet.of("tsv", "txt", "tabular", "csv", "tab", TREE_EXT);
 		final AnalysisOutputFile aof = analysis.getAnalysisOutputFile(outputName);
 		final Long aofId = aof.getId();
 		final String aofFilename = aof.getFile()
@@ -670,7 +678,7 @@ public class AnalysisController {
 	// ************************************************************************************************
 
 	/**
-	 * Construct the model parameters for an {@link AnalysisType#PHYLOGENOMICS}
+	 * Construct the model parameters for an {@link AnalysisType#PHYLOGENOMICS} or {@link AnalysisType#MLST_MENTALIST}
 	 * {@link Analysis}
 	 *
 	 * @param submission The analysis submission
@@ -682,12 +690,25 @@ public class AnalysisController {
 
 		Analysis analysis = submission.getAnalysis();
 		AnalysisOutputFile file = analysis.getAnalysisOutputFile(treeFileKey);
+		if (file == null) {
+			throw new IOException("No tree file for analysis: " + submission);
+		}
 		List<String> lines = Files.readAllLines(file.getFile());
 		model.addAttribute("analysis", analysis);
-		model.addAttribute("newick", lines.get(0));
 
-		// inform the view to display the tree preview
-		model.addAttribute("preview", "tree");
+		if (lines.size() > 1) {
+			logger.warn("Multiple lines in tree file, will only display first tree. For analysis: " + submission);
+		} else {
+			String tree = lines.get(0);
+			if (EMPTY_TREE.equals(tree)) {
+				logger.debug("Empty tree found, will hide tree preview. For analysis: " + submission);
+			} else {
+				model.addAttribute("newick", tree);
+
+				// inform the view to display the tree preview
+				model.addAttribute("preview", "tree");
+			}
+		}
 	}
 
 	/**