+++ src/main/java/ca/corefacility/bioinformatics/irida/service/workflow/IridaWorkflowsService.java
@@ -2,11 +2,7 @@ package ca.corefacility.bioinformatics.irida.service.workflow;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
-import java.util.UUID;
+import java.util.*;
 import java.util.stream.Collectors;
 
 import org.slf4j.Logger;
@@ -16,11 +12,14 @@ import org.springframework.stereotype.Service;
 
 import ca.corefacility.bioinformatics.irida.exceptions.IridaWorkflowDefaultException;
 import ca.corefacility.bioinformatics.irida.exceptions.IridaWorkflowException;
+import ca.corefacility.bioinformatics.irida.exceptions.IridaWorkflowNotDisplayableException;
 import ca.corefacility.bioinformatics.irida.exceptions.IridaWorkflowNotFoundException;
 import ca.corefacility.bioinformatics.irida.model.enums.AnalysisType;
+import ca.corefacility.bioinformatics.irida.model.enums.config.AnalysisTypeSet;
 import ca.corefacility.bioinformatics.irida.model.workflow.IridaWorkflow;
 import ca.corefacility.bioinformatics.irida.model.workflow.config.IridaWorkflowIdSet;
 import ca.corefacility.bioinformatics.irida.model.workflow.config.IridaWorkflowSet;
+import ca.corefacility.bioinformatics.irida.model.workflow.description.IridaWorkflowOutput;
 
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
@@ -43,7 +42,9 @@ public class IridaWorkflowsService {
 	 * Stores the id of a default workflow for an analysis type.
 	 */
 	private Map<AnalysisType, UUID> defaultWorkflowForAnalysis;
-
+	
+	private AnalysisTypeSet disabledAnalysisTypes;
+	
 	/**
 	 * Builds a new {@link IridaWorkflowsService} for loading up installed
 	 * workflows.
@@ -57,8 +58,28 @@ public class IridaWorkflowsService {
 	 *             If there was an issue when attempting to register the
 	 *             workflows.
 	 */
-	@Autowired
 	public IridaWorkflowsService(IridaWorkflowSet iridaWorkflows, IridaWorkflowIdSet defaultIridaWorkflows)
+			throws IridaWorkflowException {
+		this(iridaWorkflows, defaultIridaWorkflows, new AnalysisTypeSet());
+	}
+
+	/**
+	 * Builds a new {@link IridaWorkflowsService} for loading up installed
+	 * workflows.
+	 * 
+	 * @param iridaWorkflows
+	 *            A {@link IridaWorkflowSet} of {@link IridaWorkflow}s to use in IRIDA.
+	 * @param defaultIridaWorkflows
+	 *            A {@link IridaWorkflowIdSet} of {@link UUID}s to use as the default
+	 *            workflows.
+	 * @param disabledAnalysisTypes
+	 *            A {@link Set} of disabled {@link AnalysisType}s.
+	 * @throws IridaWorkflowException
+	 *             If there was an issue when attempting to register the
+	 *             workflows.
+	 */
+	@Autowired
+	public IridaWorkflowsService(IridaWorkflowSet iridaWorkflows, IridaWorkflowIdSet defaultIridaWorkflows, AnalysisTypeSet disabledAnalysisTypes)
 			throws IridaWorkflowException {
 		checkNotNull(iridaWorkflows, "iridaWorkflows is null");
 		checkNotNull(defaultIridaWorkflows, "defaultWorkflows is null");
@@ -68,6 +89,8 @@ public class IridaWorkflowsService {
 
 		registerWorkflows(iridaWorkflows.getIridaWorkflows());
 		setDefaultWorkflows(defaultIridaWorkflows.getIridaWorkflowIds());
+		
+		this.disabledAnalysisTypes = disabledAnalysisTypes;
 	}
 
 	/**
@@ -236,6 +259,15 @@ public class IridaWorkflowsService {
 
 		return types;
 	}
+	
+	/**
+	 * Gets a {@link Set} of disabled {@link AnalysisType}s.
+	 * 
+	 * @return A {@link Set} of disabled {@link AnalysisType}s.
+	 */
+	public Set<AnalysisType> getDisplayableWorkflowTypes() {
+		return Sets.difference(getRegisteredWorkflowTypes(), disabledAnalysisTypes.getAnalysisTypes());
+	}
 
 	/**
 	 * Returns a workflow with the given id.
@@ -254,6 +286,42 @@ public class IridaWorkflowsService {
 		} else {
 			throw new IridaWorkflowNotFoundException(workflowId);
 		}
+	}
+	
+	/**
+	 * Returns a workflow with the given id that is displayable.
+	 * 
+	 * @param workflowId The identifier of the workflow to get.
+	 * @return An {@link IridaWorkflow} with the given identifier that is
+	 *         displayable.
+	 * @throws IridaWorkflowNotDisplayableException If no workflow with the given
+	 *                                              identifier is not displayable.
+	 * @throws IridaWorkflowNotFoundException       If the workflow was not found.
+	 */
+	public IridaWorkflow getDisplayableIridaWorkflow(UUID workflowId)
+			throws IridaWorkflowNotDisplayableException, IridaWorkflowNotFoundException {
+		IridaWorkflow workflow = getIridaWorkflow(workflowId);
+
+		if (getDisplayableWorkflowTypes().contains(workflow.getWorkflowDescription().getAnalysisType())) {
+			return workflow;
+		} else {
+			throw new IridaWorkflowNotDisplayableException(workflowId);
+		}
+	}
+
+	/**
+	 * Get list of workflow output names.
+	 * @param workflowId Workflow UUID.
+	 * @return List of workflow output names.
+	 * @throws IridaWorkflowNotFoundException if no workflow with the given UUID found.
+	 */
+	public List<String> getOutputNames(UUID workflowId) throws IridaWorkflowNotFoundException {
+		final IridaWorkflow iridaWorkflow = getIridaWorkflow(workflowId);
+		return iridaWorkflow.getWorkflowDescription()
+				.getOutputs()
+				.stream()
+				.map(IridaWorkflowOutput::getName)
+				.collect(Collectors.toList());
 	}
 
 	/**